google-adk/requirements.txt
google-adk
sqlalchemy
python-dotenv